// src/app/pages/model/Profile.ts
export class Profile {
  id: number | undefined;
  apiProfileId: number | undefined;
  companyName: string | undefined;
  cisNumber: number | undefined;
  dcarNumber: number | undefined;
  registrationNumber: number | undefined;
  profileStatus: string | undefined;
  prefix: string | undefined;
  profileState: string | undefined;
  clientType: string | undefined;
  productType: string | undefined;
  connectDirectFileType: string | undefined;
  marketSector: string | undefined;
  connectDirectAdd: string | undefined;
  insertDate: string | undefined;
  updateDate: string | undefined;
 
}